feat(event_handler): support richer Tags on OpenAPI object
